xilinx: board: Add support for additional card detection
The most of Xilinx evaluation boards have FMC connectors which contain
small eeprom for card identification. That's why read content of eeprom and
record it.
Also generate cardX_ variables for easier script handling.

Signed-off-by: Michal Simek <michal.simek@xilinx.com>
Series-to: uboot
Cover-letter:
xilinx: board: Add support for board eeproms
Hi,

Based on discussion with Rob we should be using nvmem alias to pointing to
eeprom which stores information about device. ZynqMP boards are using
format which is based on offset. The code is reading it and setup
environment variables based on it. They can be used within scripts to
automate different behavior for specific boards.

Thanks,
Michal
END